requirements.txt
# Deep Learning & Machine Learning
tensorflow==2.16.2
scikit-learn==1.6.0

# Image Processing
Pillow==10.4.0
opencv-python==4.10.0.84
pytesseract==0.3.13
pdf2image==1.17.0

# Scientific Computing
numpy==1.26.4